==> 12.RemoveFunctions.R <==

library(openxlsx)

?openxlsx::removeCellMerge

?openxlsx::removeColWidths
?openxlsx::removeRowHeights

?openxlsx::removeComment

?openxlsx::removeFilter

?openxlsx::removeTable

?openxlsx::removeWorksheet
